Renamed file to avoid duplicate file naming.
The performance test framework consists of a framework running on the test
server and one that runs on client devices. This commit renames the client-side
file to PerfTestFrameworkClient.js to distinguish from PerfTestFramework.js,
which is the one used on the server.
