+++ test/www/jxcore/perf_tests/PerfTestFrameworkClient.js
